agent_example/requirements.txt
transformers
torch
openai
python-dotenv
google-search-results
autogen-agentchat
autogen-ext
tiktoken
